LoLDB sql file and php files
This code is the front end of the website and an sql file with some
tables in it for functionalities sake. This front end currently has the
capabilities checking the database to see if a user already exists,
checking the users password for correctness, and adding a new user if
one does not yet exist and inserting that user into the database table
"user". Users who already exist can view their (or other players)
"stats" table.