logout, session destroy, and consistent sessions added
updated session stars and session names in files to be consistent
amongst all files so that correct user is delayed and errors are not
thrown. also added a logout functionality that allows for sessions to be
destroyed and new sessions to begin on the same browser without deleting
cookies or session information in the browser settings. began work on
create new user so that an entry in the stats table is made upon
creation of a new user. this can be simplified by using a stored
procedure in MySQL.