stored procedure
above utilizes the stored procedure sp_createUser. I update leaguedb's
stored procedure to work correctly and got it up and running. This front
end is essentially ready to be used however we wish, it just needs some
CSS to make it look good and a backend that is better developed! :)